requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.5
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.4.1
django-appconf==1.1.0
django-cors-headers==4.7.0
django-fernet-fields==0.6
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
gunicorn==23.0.0
honcho==2.0.0
idna==3.10
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.2.0
requests==2.32.3
six==1.17.0
slack_sdk==3.35.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
